requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Deprecated==1.2.18
frozenlist==1.5.0
get-annotations==0.1.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.4
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
yandex-cloud-ml-sdk==0.8.0
yandexcloud==0.337.0
yarl==1.18.3